Cleanup and consolidate types
